requirements.txt
Flask
pdfminer.six
python-docx
nltk
scikit-learn
sentence-transformers
torch
